Moved all the JS to browserify
