+++ dev/_assets/scripts/components/helpers.js
@@ -1,0 +1,10 @@
+function extend( a, b ) {
+    for( var key in b ) {
+        if( b.hasOwnProperty( key ) ) {
+            a[key] = b[key];
+        }
+    }
+    return a;
+}
+
+module.exports = extend;
