+++ dev/_assets/scripts/json/identities.json
@@ -1,0 +1,40 @@
+[{
+    "mce-NAME": "Luke Skywalker",
+    "mce-EMAIL": "luke@rebelalliance.org"
+},{
+    "mce-NAME": "Jake",
+    "mce-EMAIL": "the_dog@adventureti.me"
+},{
+    "mce-NAME": "Finn",
+    "mce-EMAIL": "the_human@adventureti.me"
+},{
+    "mce-NAME": "Ron Burgundy",
+    "mce-EMAIL": "r.burgundy@channel4.tv"
+},{
+    "mce-NAME": "Batman",
+    "mce-EMAIL": "bruce@wayne.org"
+},{
+    "mce-NAME": "Doc Brown",
+    "mce-EMAIL": "emmet.brown@hillvalley.edu"
+},{
+    "mce-NAME": "Walter White",
+    "mce-EMAIL": "heisenberg@savewalterwhite.com"
+},{
+    "mce-NAME": "Charlie Kelly",
+    "mce-EMAIL": "teh_niteman@hotmale.com"
+},{
+    "mce-NAME": "Jeff Lebowski",
+    "mce-EMAIL": "the-dude@gmail.com"
+},{
+    "mce-NAME": "Peter",
+    "mce-EMAIL": "p.gibbins@initech.com"
+},{
+    "mce-NAME": "Stanley Ipkiss",
+    "mce-EMAIL": "somebody-stop-me@yahoo.com"
+},{
+    "mce-NAME": "Pete Mitchell",
+    "mce-EMAIL": "maverick@topgunacademy.gov"
+},{
+    "mce-NAME": "Rimmer",
+    "mce-EMAIL": "a.rimmer@jmc.org"
+}]
