+++ grunt/uglify.js
@@ -1,7 +1,7 @@
 module.exports = {
     scripts: {
         files: {
-            'web/scripts/main.js': ['dev/_assets/scripts/*.js']
+            'web/scripts/main.js': ['web/scripts/main.js']
         }
     },
 };
