+++ package.json
@@ -5,6 +5,7 @@
   "devDependencies": {
     "grunt": "^0.4.5",
     "grunt-autoprefixer": "^1.0.1",
+    "grunt-browserify": "^3.3.0",
     "grunt-contrib-clean": "^0.6.0",
     "grunt-contrib-connect": "^0.7.1",
     "grunt-contrib-copy": "^0.7.0",
